requirements.txt
plyfile
tqdm
scipy
opencv-python
matplotlib
imageio
imageio-ffmpeg
lpips
pytorch-msssim
torchmetrics
omegaconf
tensorboard
wandb
pydantic
torchdiffeq
torchode
torchtyping
functorch
dearpygui
gdown
gitpython
ipykernel
beautifulsoup4
numpy==1.25.2
pillow
requests
ninja
cmake